Set permissions on volumes defined in the Dockerfile
VOLUMES defined in the Dockerfile will get root:root ownership in the
container. This will lead to the following message on startup:
"Cannot write to content dirs('/var/lib/zoneminder/events',
'/var/lib/zoneminder/images'). Check that these exist and are
owned by the web account user".

I added the permission changing `chown` to the entrypoint script. I
did this in the entrypoint, and not the build, because even if this was
set in build, if a user bind mounts a volume (definitely docker-compose;
not sure about just docker), it will end up as root anyway. Best to do
in the entrypoint.